Update GitHub Actions workflows and optimize caching
restore input option for branch checkout in GitHub
setup-and-build-nocheck action

Update GitHub Actions workflows

All GitHub Actions workflows were updated to include the "Checkout sources" step. Moreover, caching has been improved in the Rust workflow and the timeout limit was adjusted in the prover-test workflow. Optional usage of submodules during checkout was also removed in some workflows.

Update GitHub Actions to include Rust cache and Solana keypair generation

Remove specific pnpm version from setup in workflows

Update cache paths in GitHub actions

Add disk space usage to lint workflow

Update cache paths in GitHub actions

Update GitHub action setup and cache management

Update cache paths in GitHub actions

Update caching strategy in GitHub actions

Remove build step from setup-and-build action

Enable pnpm installation in setup-and-build action

Move '@lightprotocol/hasher.rs' to dependencies

The '@lightprotocol/hasher.rs' was wrongly placed in the devDependencies section in package.json. With this commit, it has been moved to the dependencies section to ensure its availability for production runtime.

Update pnpm-lock.yaml dependencies

The '@lightprotocol/hasher.rs' package has been moved within the pnpm-lock.yaml file, indicating a change in the dependency structure. This rearrangement might be due to a modification in the project's build order or a dependency update.

Add '@lightprotocol/zk-compression-cli' to devDependencies

The '@lightprotocol/zk-compression-cli' package has been included as a new development dependency in the circuit-lib.js/package.json file. Corresponding updates were also made in the pnpm-lock.yaml file to reflect this change and ensure package version consistency.

Add zk-compression-cli to forester dependencies

A new dependency, '@lightprotocol/zk-compression-cli', has been added to both pnpm-lock.yaml and package.json files.

Add .gnark-prover/proving-keys to GitHub actions cache

The commit enhances caching in our GitHub actions CI workflow. The .gnark-prover/proving-keys directory has been added to the paths being cached. This is expected to accelerate build times by reusing the proving keys between runs.

Update test commands in GitHub workflow

The test commands in the GitHub workflow for 'stateless.js' and 'compressed-token' have been revised. Instead of using the previous 'cd' and 'pnpm' commands, 'npx nx' is now used for a more direct and efficient test run sequence.

Add '@lightprotocol/zk-compression-cli' to dependencies

The '@lightprotocol/zk-compression-cli' has been added to the 'devDependencies' in 'package.json' and to the list of dependencies in 'pnpm-lock.yaml'. Also, the build of this package has been included in the GitHub Actions workflow. Besides, the comment in 'install.sh' script has been updated to reflect the correct directory for placing gnark keys.

Remove '@lightprotocol/zk-compression-cli' from dependencies

The dependency '@lightprotocol/zk-compression-cli' is removed from both 'package.json' and 'pnpm-lock.yaml' files. This change indicates that this particular package is no longer necessary in the current implementation of the project.

Use nx/test in forester workflow

Upgrade GitHub Actions and adjust cache paths

Updated the use of GitHub Actions from version 3 to version 4 to leverage the latest improvements and functionalities. Also modified the cache paths for gnark-prover, removing the leading dot to reflect the correct directory structure.

Refactor GitHub Action workflows and centralize setup tasks

The existing workflows have been refactored to use a centralized GitHub Action for most setup tasks, promoting consistent behaviour across workflows. Some redundant steps, such as manually copying 'spl_noop.so' in multiple workflows, have been moved into the shared 'setup-and-build' action. Additionally, the caching strategy has been simplified by removing manual dependency installation in favor of the unified setup action.

Add build step for CLI in workflow

A new step has been added to the Github Actions workflow in the 'light-system-programs-tests.yml' file. This specific step is responsible for building the CLI by running the '@lightprotocol/zk-compression-cli:build' command.

Update build command in workflow files

Track 'main' branch in pull requests

Added a command to track the 'main' branch when the event is a pull request in both Light System Programs and JS workflows.

Modify GitHub Action workflows to fetch full history

The changes have updated the `light

Update nx-cloud version to 19.0.0

The nx-cloud version has been updated from 16.5.2 to 19.0.0 in package.json, pnpm-lock.yaml, and nx

Upgrade nx version to 19.3.0

The nx package has been upgraded from version 16.10.0 to 19.3.0. This change will incorporate any updates, bug fixes, and improvements made since the last version. All relevant references and dependencies have been updated accordingly.

Update program tags for Light system modules

Program tags for Light system modules such as 'light_system_program', 'light_compressed_token', 'account_compression', and 'light_registry' have been set to reflect their appropriate module tags in 'initTestEnv.ts'. Related constants have also been updated in 'constants.ts'.

Add output paths to nx.json configuration file

New output paths have been added to the nx.json file to configure where built projects are stored. These additional paths target deploy and idl, allowing for a more organized and clear structure for generated output.

Add new output path in nx.json

A new output path "{workspaceRoot}/target/types" has been added in nx.json. This path will be used for depositing the compiled typescript files.

Update build script and nx.json configuration

The build script has been simplified by removing the "--skip-nx-cache" flag from the 'npx nx run-many' command usage. In the nx.json configuration, caching has been enabled for the build and test targets. Additionally, a new "noTestLedger" named input has been introduced to exclude certain paths during build and test processes.

Add nx targets and update scripts, remove nx-cloud

This commit adds 'nx' targets to several package.json files, creating a structure for running and building projects with nx. The 'nx-cloud' dependency has been removed, which also entails removing its configurations from the 'tasksRunnerOptions' in nx.json. Some scripts and ignore file entries have been updated accordingly.

Remove debug output from build targets

+++ scripts/build.sh
@@ -9,8 +9,10 @@ set -eux
 
 pnpm install || { echo >&2 "Failed to install dependencies. Aborting."; exit 1; }
 
-mkdir -p target/deploy && cp third-party/solana-program-library/spl_noop.so target/deploy
+if [ ! -f target/deploy/spl_noop.so ]; then
+    mkdir -p target/deploy && cp third-party/solana-program-library/spl_noop.so target/deploy
+fi
 
-npx nx run-many --target=build --all --skip-nx-cache
+npx nx run-many --target=build --all
 
 echo "Build process completed successfully."
